% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preset-report_wizard.R
\name{create_report_wizard}
\alias{create_report_wizard}
\title{Create report wizard to be used within the interactive modules}
\usage{
create_report_wizard(
  pipeline,
  session = shiny::getDefaultReactiveDomain(),
  env = parent.frame()
)
}
\arguments{
\item{pipeline}{\code{ravepipeline} pipeline}

\item{session}{shiny session}

\item{env}{environment used to register the wizard inputs with
\pkg{shidashi}, so that agents can choose and generate reports; default is
the calling frame, typically the module server}
}
\value{
A list of functions: \code{launch} with argument \code{subject} to
be called when users want to pop up a wizard allowing users to choose
reports; \code{generate} with arguments \code{subject} and
\code{report_names} to generate reports
}
\description{
Create report wizard to be used within the interactive modules
}
