% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rave-launch.R
\name{debug_modules}
\alias{debug_modules}
\title{Debug 'RAVE' modules interactively in local project folder}
\usage{
debug_modules(
  module_root = rstudioapi::getActiveProject(),
  host = "127.0.0.1",
  port = 17283,
  jupyter = FALSE,
  ...
)
}
\arguments{
\item{module_root}{root of modules, usually the project folder created from
\code{'shidashi'} template}

\item{host, port}{host and port of the application}

\item{jupyter}{whether to launch \code{'Jupyter'} server; default is false}

\item{...}{passed to \code{\link[shidashi]{render}}}
}
\value{
\code{'RStudio'} job ID
}
\description{
Debug 'RAVE' modules interactively in local project folder
}
