% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-utils.R
\name{get_active_module_info}
\alias{get_active_module_info}
\title{Get current active module information, internally used}
\usage{
get_active_module_info(session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{session}{shiny reactive domain, default is current domain}
}
\value{
A named list, including module ID, module label, internal
\code{'rave_id'}.
}
\description{
Get current active module information, internally used
}
