% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-new_data.R
\name{is_new_repository_loaded}
\alias{is_new_repository_loaded}
\title{Check whether repository read from pipeline has been updated}
\usage{
is_new_repository_loaded(
  component_container,
  pipeline,
  repository_class = NA,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{component_container}{A \code{'RAVEShinyComponentContainer'} instance}

\item{pipeline}{A \code{\link[ravepipeline]{pipeline}} instance}

\item{repository_class}{Expected repository class; default is arbitrary}

\item{session}{A shiny session}
}
\value{
Logical \code{TRUE} or \code{FALSE} is new repository is loaded
}
\description{
Helper function to check if new pipeline repository is loaded; see 'Examples'
}
\examples{


\dontrun{

shiny::bindEvent(
  ravedash::safe_observe({
    if (!ravedash::on_new_repository_loaded()) {
      return()
    }

    # Handle event when new repository is loaded

  }, priority = 1001),
  ravedash::watch_data_loaded(),
  ignoreNULL = FALSE,
  ignoreInit = FALSE
)


}



}
