% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-utils.R
\name{load_data_button}
\alias{load_data_button}
\title{Button to trigger data loader}
\usage{
load_data_button(
  label = "Load data",
  icon = NULL,
  width = NULL,
  type = "primary",
  btn_type = c("button", "link"),
  class = "",
  style = "",
  ...
)
}
\arguments{
\item{label}{label to display}

\item{icon}{icon before the label}

\item{width, class, style, ...}{passed to 'HTML' tag}

\item{type}{used to calculate \code{class}}

\item{btn_type}{button style, choices are \code{'button'} or \code{'link'}}
}
\value{
A 'HTML' button tag
}
\description{
A button that triggers \code{'load_data'} event, which runs
scripts registered via \code{set_script} with
\code{binding_event = "load_data"} (see \code{\link{module_server_common}});
the same script can be triggered programmatically via
\code{trigger_script}.
}
\examples{

# In loader UI
load_data_button("Load subject", width = "100\%")

# In loader server function
\dontrun{
server_tools <- get_default_handlers(session = session)
server_tools$set_script(
  "load_data",
  {
    # load data ...
  },
  binding_event = "load_data",
  dispatch_event = "data_changed",
  alert_params = list(title = "Loading in progress")
)
}

}
\seealso{
\code{\link{run_analysis_button}}, \code{\link{get_rave_event}}
}
