% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\name{obsolete}
\alias{obsolete}
\alias{output_gadget_container}
\title{Obsolete functions}
\usage{
output_gadget_container(expr, ..., quoted = FALSE, env = parent.frame())
}
\arguments{
\item{expr}{expression to evaluate}

\item{...}{ignored}

\item{quoted}{whether \code{expr} is quoted}

\item{env}{environment to evaluate \code{expr}}
}
\description{
Reserved for compatibility; do not use.
}
