% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output-widgets.R
\name{output_gadget}
\alias{output_gadget}
\alias{output_gadget_container}
\title{'RAVE' dashboard output gadgets}
\usage{
output_gadget(
  outputId,
  icon = NULL,
  type = c("standalone", "download", "actionbutton", "custom"),
  class = NULL,
  inputId = NULL,
  ...
)

output_gadget_container(
  expr,
  gadgets = c("standalone", "download"),
  quoted = FALSE,
  env = parent.frame(),
  outputId = NULL,
  class = NULL,
  container = NULL,
  wrapper = TRUE
)
}
\arguments{
\item{outputId}{output ID in the root scope of shiny session}

\item{icon}{gadget icon}

\item{type, gadgets}{gadget type(s), currently supported: \code{'standalone'},
\code{'download'}, \code{'actionbutton'}}

\item{class}{additional class to the gadget or its container}

\item{inputId}{input ID, automatically assigned internally}

\item{...}{ignored}

\item{expr}{shiny output call expression, for example,
\code{shiny::plotOutput({...})}}

\item{quoted}{whether \code{expr} is quoted; default is false}

\item{env}{environment where \code{expr} should be evaluated}

\item{container}{optional container for the gadgets and outputs; will be ignored
if \code{wrapper} is false}

\item{wrapper}{whether to wrap the gadgets and the output within
a 'HTML' container}
}
\description{
'RAVE' dashboard output gadgets
}
