% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-outputs.R
\name{plotOutput2}
\alias{plotOutput2}
\title{Shiny plot output with minimum height and additional classes}
\usage{
plotOutput2(
  outputId,
  class = NULL,
  width = "100\%",
  height = "100\%",
  min_height = "400px",
  ...
)
}
\arguments{
\item{outputId, width, height, ...}{passed to \code{\link[shiny]{plotOutput}}}

\item{class}{additional 'HTML' class of the output wrapper}

\item{min_height}{minimum height of the image; default is 400 pixels}
}
\value{
A plot output element that can be included in a panel.
}
\description{
Shiny plot output with minimum height and additional classes
}
\examples{

plotOutput2("plot", class = "rounded overflow-hidden",
            min_height = 300)


}
