% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{random-text}
\alias{random-text}
\alias{be_patient_text}
\alias{finished_text}
\title{Randomly choose a text from a list of strings}
\usage{
be_patient_text(candidates)

finished_text(candidates)
}
\arguments{
\item{candidates}{character vectors, a list of candidates}
}
\value{
\code{be_patient_text} returns a text asking users to be patient;
\code{finished_text} returns the text indicating the task has finished.
}
\description{
Randomly choose a text from a list of strings
}
\examples{

be_patient_text()

finished_text()

}
