% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-utils.R
\name{ravedash_footer}
\alias{ravedash_footer}
\title{A hovering footer at bottom-right}
\usage{
ravedash_footer(module_id = NULL)
}
\arguments{
\item{module_id}{'RAVE' module ID}
}
\value{
'HTML' tags
}
\description{
The footer provides buttons to open/close the loader panel,
show brief message of data loaded, and links to jump to input cards
}
\examples{

library(shiny)
# dummy variables for the example
data_loaded <- TRUE

# UI code
ravedash_footer("my_module")

# server code to set message
server <- function(input, output, session){

  module_server_common(input, output, session, function(){

    # check if data has been loaded
    if(data_loaded) {

      # if yes, then set the footer message
      fire_rave_event("loader_message",
                      "my_project/subject - Epoch: Auditory")
      return(TRUE)
    } else {

      # No data found, unset the footer message
      fire_rave_event("loader_message", NULL)
      return(FALSE)
    }

  })
}

}
