% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-utils.R
\name{shiny_cache}
\alias{shiny_cache}
\title{Obtain caching object for current run-time shiny session}
\usage{
shiny_cache(namespace, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{namespace}{characters, usually the module ID}

\item{session}{shiny interactive context domain}
}
\value{
A caching object. The caching object is identical within the same
context and namespace.
}
\description{
Cache small objects such as inputs or configurations
}
