% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-icons.R
\name{shiny_icons}
\alias{shiny_icons}
\title{Shiny icons}
\usage{
shiny_icons
}
\description{
Shiny icons
}
\details{
The goal of create this list is to keep 'shiny' icons (which are
essentially 'font-awesome' icons) up-to-date.
}
