% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-common.R
\name{switch_module}
\alias{switch_module}
\title{Drive 'RAVE' browser to switch to another module}
\usage{
switch_module(module_id, title, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{module_id}{the module ID, see \code{'modules.yaml'} in the pipeline
directory}

\item{title}{the module title to display}

\item{session}{shiny session}
}
\value{
Nothing
}
\description{
Switch to another 'RAVE' module to continue the procedures.
}
