% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preset-report_wizard.R
\name{create_report_wizard}
\alias{create_report_wizard}
\title{Create report wizard}
\usage{
create_report_wizard(pipeline, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{pipeline}{\code{ravepipeline} pipeline}

\item{session}{shiny session}
}
\value{
A function with argument \code{subject}; this function is to be
called when users want to create reports, must be used within the module
server code.
}
\description{
Create report wizard
}
