% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preset-report_wizard.R
\name{create_report_wizard}
\alias{create_report_wizard}
\title{Create report wizard to be used within the interactive modules}
\usage{
create_report_wizard(pipeline, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{pipeline}{\code{ravepipeline} pipeline}

\item{session}{shiny session}
}
\value{
A list of functions: \code{launch} with argument \code{subject} to
be called when users want to pop up a wizard allowing users to choose
reports; \code{generate} with arguments \code{subject} and
\code{report_names} to generate reports
}
\description{
Create report wizard to be used within the interactive modules
}
