% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preset-loader-3dviewer.R,
%   R/preset-loader-electrode.R, R/preset-loader-epoch.R,
%   R/preset-loader-project.R, R/preset-loader-reference.R,
%   R/preset-loader-subject.R
\name{presets_loader_3dviewer}
\alias{presets_loader_3dviewer}
\alias{presets_loader_electrodes}
\alias{presets_loader_epoch}
\alias{rave-ui-preset}
\alias{presets_loader_project}
\alias{presets_loader_reference}
\alias{presets_loader_subject}
\title{Preset reusable 'RAVE' front-end elements}
\usage{
presets_loader_3dviewer(
  id = "loader_3d_viewer",
  height = "600px",
  loader_project_id = "loader_project_name",
  loader_subject_id = "loader_subject_code",
  loader_reference_id = "loader_reference_name",
  loader_electrodes_id = "loader_electrode_text"
)

presets_loader_electrodes(
  id = "loader_electrode_text",
  varname = "loaded_electrodes",
  label = "Electrodes",
  loader_project_id = "loader_project_name",
  loader_subject_id = "loader_subject_code"
)

presets_loader_epoch(
  id = "loader_epoch_name",
  varname = "epoch_choice",
  label = "Epoch and Trial Duration",
  loader_project_id = "loader_project_name",
  loader_subject_id = "loader_subject_code"
)

presets_loader_project(
  id = "loader_project_name",
  varname = "project_name",
  label = "Project"
)

presets_loader_reference(
  id = "loader_reference_name",
  varname = "reference_name",
  label = "Reference name",
  loader_project_id = "loader_project_name",
  loader_subject_id = "loader_subject_code"
)

presets_loader_subject(
  id = "loader_subject_code",
  varname = "subject_code",
  label = "Subject",
  loader_project_id = "loader_project_name"
)
}
\arguments{
\item{id}{input or output ID of the element; this ID will be prepended with
module namespace}

\item{height}{height of the element}

\item{loader_project_id}{the ID of \code{presets_loader_project} if
different to the default}

\item{loader_subject_id}{the ID of \code{presets_loader_subject} if
different to the default}

\item{loader_reference_id}{the ID of \code{presets_loader_reference} if
different to the default}

\item{loader_electrodes_id}{the ID of \code{presets_loader_electrodes} if
different to the default}

\item{varname, varnames}{variable name(s) in the module's settings file}

\item{label, labels}{readable label(s) of the element}
}
\description{
Preset reusable 'RAVE' front-end elements
}
