% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-rave-comps.R
\name{rave-ui-preset}
\alias{rave-ui-preset}
\alias{presets_loader_project}
\alias{presets_loader_subject}
\alias{presets_loader_epoch}
\alias{presets_loader_reference}
\alias{presets_loader_electrodes}
\alias{presets_loader_3dviewer}
\title{Preset reusable 'RAVE' front-end elements}
\usage{
presets_loader_project(
  id = "loader_project_name",
  varname = "project_name",
  label = "Project"
)

presets_loader_subject(
  id = "loader_subject_code",
  varname = "subject_code",
  label = "Subject",
  loader_project_id = "loader_project_name"
)

presets_loader_epoch(
  id = "loader_epoch_name",
  varnames = c("epoch_name", "trial_starts", "trial_ends"),
  labels = c("Epoch name", "Pre", "Post"),
  loader_project_id = "loader_project_name",
  loader_subject_id = "loader_subject_code"
)

presets_loader_reference(
  id = "loader_reference_name",
  varname = "reference_name",
  label = "Reference name",
  loader_project_id = "loader_project_name",
  loader_subject_id = "loader_subject_code"
)

presets_loader_electrodes(
  id = "loader_electrode_text",
  varname = "loaded_electrodes",
  label = "Electrodes",
  loader_project_id = "loader_project_name",
  loader_subject_id = "loader_subject_code"
)

presets_loader_3dviewer(
  id = "loader_3d_viewer",
  height = "600px",
  loader_project_id = "loader_project_name",
  loader_subject_id = "loader_subject_code",
  loader_reference_id = "loader_reference_name",
  loader_electrodes_id = "loader_electrode_text"
)
}
\arguments{
\item{id}{input or output ID of the element; this ID will be prepended with
module namespace}

\item{varname, varnames}{variable name(s) in the module's settings file}

\item{label, labels}{readable label(s) of the element}

\item{loader_project_id}{the ID of \code{presets_loader_project} if
different to the default}

\item{loader_subject_id}{the ID of \code{presets_loader_subject} if
different to the default}

\item{height}{height of the element}

\item{loader_reference_id}{the ID of \code{presets_loader_reference} if
different to the default}

\item{loader_electrodes_id}{the ID of \code{presets_loader_electrodes} if
different to the default}
}
\description{
Preset reusable 'RAVE' front-end elements
}
