% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%OF\%}
\alias{shiny_alert2}
\alias{close_alert2}
\alias{show_notification}
\alias{clear_notifications}
\alias{register_output}
\alias{card_badge}
\alias{card_recalculate_badge}
\alias{enable_recalculate_badge}
\alias{disable_recalculate_badge}
\alias{set_card_badge}
\alias{logger}
\alias{set_logger_path}
\alias{logger_threshold}
\alias{logger_error_condition}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dipsaus}{\code{\link[dipsaus:grapes-OF-grapes]{\%OF\%}}, \code{\link[dipsaus:shiny_alert2]{close_alert2}}, \code{\link[dipsaus]{shiny_alert2}}}

  \item{ravepipeline}{\code{\link[ravepipeline]{logger}}, \code{\link[ravepipeline:logger]{logger_error_condition}}, \code{\link[ravepipeline:logger]{logger_threshold}}, \code{\link[ravepipeline:logger]{set_logger_path}}}

  \item{shidashi}{\code{\link[shidashi]{card_badge}}, \code{\link[shidashi:card_badge]{card_recalculate_badge}}, \code{\link[shidashi:notification]{clear_notifications}}, \code{\link[shidashi:card_badge]{disable_recalculate_badge}}, \code{\link[shidashi:card_badge]{enable_recalculate_badge}}, \code{\link[shidashi:register_io]{register_output}}, \code{\link[shidashi:card_badge]{set_card_badge}}, \code{\link[shidashi:notification]{show_notification}}}
}}

