% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-utils.R
\name{run_analysis_button}
\alias{run_analysis_button}
\title{Button to trigger analysis}
\usage{
run_analysis_button(
  label = "Run analysis (Ctrl+Enter)",
  icon = NULL,
  width = NULL,
  type = "primary",
  btn_type = "button",
  class = "",
  ...
)
}
\arguments{
\item{label}{label to display}

\item{icon}{icon before the label}

\item{width, btn_type, class, ...}{passed to 'HTML' \code{button} tag}

\item{type}{used to calculate \code{class}}
}
\value{
A 'HTML' button tag
}
\description{
A button that triggers \code{'run_analysis'} event;
see also \code{\link{get_rave_event}}
}
