% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-utils.R
\docType{data}
\name{shiny_icons}
\alias{shiny_icons}
\title{Shiny icons}
\format{
An object of class \code{ravedash_shiny_icons} of length 0.
}
\usage{
shiny_icons
}
\description{
Shiny icons
}
\details{
The goal of create this list is to keep 'shiny' icons (which are
essentially 'font-awesome' icons) up-to-date.
}
\keyword{datasets}
