% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{with_error_notification}
\alias{with_error_notification}
\alias{error_notification}
\alias{error_alert}
\alias{with_error_alert}
\title{Evaluate with automatic error handlers in dashboard}
\usage{
error_notification(
  cond,
  title = "Error found!",
  type = "danger",
  class = "error_notif",
  delay = 30000,
  autohide = TRUE,
  collapse = "\\n",
  prefix = paste("Found the following error",
    "(details have been printed in the console):"),
  session = shiny::getDefaultReactiveDomain()
)

error_alert(
  cond,
  title = "Error found!",
  type = "error",
  danger_mode = TRUE,
  auto_close = FALSE,
  prefix = paste("Found the following error",
    "(details have been printed in the console):"),
  buttons = "Confirm",
  session = shiny::getDefaultReactiveDomain()
)

with_error_notification(expr, envir = parent.frame(), quoted = FALSE, ...)

with_error_alert(expr, envir = parent.frame(), quoted = FALSE, ...)
}
\arguments{
\item{cond}{condition to log}

\item{type}{which type of logging should be set; default is \code{'console'},
if file log is enabled through \code{set_logger_path}, \code{type} could be
\code{'file'} or \code{'both'}. Default log level is \code{'info'} on
console and \code{'debug'} on file.}

\item{class, title, delay, autohide}{passed to \code{\link[shidashi]{show_notification}}}

\item{collapse, danger_mode, auto_close, buttons}{will be passed to
\code{\link[dipsaus]{shiny_alert2}} or
\code{\link[shidashi]{show_notification}}}

\item{prefix}{additional messages to display in the notification or alert}

\item{session}{shiny session}

\item{expr}{expression to evaluate}

\item{envir}{environment to evaluate \code{expr}}

\item{quoted}{whether \code{expr} is quoted; default is false}

\item{...}{passed to other methods}
}
\value{
The condition \code{cond} if errored out, or the evaluated results
}
\description{
Keep track of messages printed by modules
}
